Rakesh_my/requirements.txt
flask
gunicorn
python-telegram-bot
pandas
openpyxl
python-pptx
werkzeug